move validity checks to separate file
